updated dependencies and vacation support
