+++ app/index.html
@@ -39,10 +39,10 @@
     <!-- build:js scripts/vendor.js -->
     <!-- bower:js -->
     <script src="/bower_components/jquery/dist/jquery.js"></script>
+    <script src="/bower_components/bootstrap/dist/js/bootstrap.js"></script>
     <script src="/bower_components/lodash/lodash.js"></script>
     <script src="/bower_components/moment/moment.js"></script>
     <script src="/bower_components/moment-range/dist/moment-range.js"></script>
-    <script src="/bower_components/holidays/src/Holidays.js"></script>
     <!-- endbower -->
     <!-- endbuild -->
