+++ bower.json
@@ -5,8 +5,8 @@
     "bootstrap": "~3.3.5",
     "holidays": "~0.1.0",
     "lodash": "~3.10.1",
-    "moment": "~2.10.6",
-    "moment-range": "~2.0.3"
+    "moment": "^2.14.1",
+    "moment-range": "^2.2.0"
   },
   "overrides": {
     "bootstrap": {
