+++ package.json
@@ -1,40 +1,47 @@
 {
   "private": true,
   "engines": {
-    "node": ">=0.12.0"
+    "node": ">=4"
   },
   "devDependencies": {
-    "babel-core": "^5.5.6",
-    "bower": "^1.5.2",
+    "babel-core": "^6.4.0",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-register": "^6.5.2",
     "browser-sync": "^2.2.1",
     "del": "^1.1.1",
     "gulp": "^3.9.0",
-    "gulp-autoprefixer": "^2.3.1",
-    "gulp-cache": "^0.2.8",
-    "gulp-eslint": "^0.13.2",
-    "gulp-if": "^1.2.5",
+    "gulp-autoprefixer": "^3.0.1",
+    "gulp-babel": "^6.1.1",
+    "gulp-cache": "^0.4.2",
+    "gulp-cssnano": "^2.0.0",
+    "gulp-eslint": "^2.0.0",
+    "gulp-htmlmin": "^1.3.0",
+    "gulp-if": "^2.0.0",
     "gulp-imagemin": "^2.2.1",
     "gulp-load-plugins": "^0.10.0",
-    "gulp-minify-css": "^1.1.1",
-    "gulp-minify-html": "^1.0.0",
+    "gulp-plumber": "^1.0.1",
+    "gulp-sass": "^2.0.0",
     "gulp-size": "^1.2.1",
     "gulp-sourcemaps": "^1.5.0",
-    "gulp-subtree": "^0.1.0",
     "gulp-uglify": "^1.1.0",
-    "gulp-useref": "^1.1.1",
+    "gulp-useref": "^3.0.0",
     "main-bower-files": "^2.5.0",
-    "opn": "^1.0.1",
     "wiredep": "^2.2.2"
+  },
+  "dependencies": {
+    "moment": "^2.14.1"
   },
   "eslintConfig": {
     "env": {
+      "es6": true,
+      "node": true,
       "browser": true
+    },
+    "rules": {
+      "quotes": [
+        2,
+        "single"
+      ]
     }
-  },
-  "scripts": {
-    "postinstall": "./node_modules/bower/bin/bower cache clean && ./node_modules/bower/bin/bower install"
-  },
-  "dependencies": {
-    "moment": "^2.10.6"
   }
 }
